housekeeping

+++ Polynomial.h
@@ -58,14 +58,14 @@ public:
 	// Increment/decrement overloads
 	Polynomial& operator--();
 	Polynomial& operator++();
-	int operator++(int);
+	float operator++(int);
 	
 	// Comparison operator overloads
 	bool operator==(const Polynomial&);
 	bool operator!=(const Polynomial& rVal) { return !(*this == rVal); };
 
 	// Evaluation (function operator) overload
-	int operator()(const int);
+	float operator()(const float);
 
 	// Friend functions for stream insertion/extraction
 	friend std::ostream& operator<<(std::ostream&, const Polynomial&);
